Fix VSIXInstaller build
The project file was a mix of NuGet references via project.json and hint
paths into Visual Studio. This produced an inconsistent set of
references that caused MSBuild conflict errors at build times and often
an unrunnable application.

Decided to unify on HintPath references for now. At the same time
deleted the project.json file and moved to PackageReference.