Stop GH merge bot from closing existing PRs
The merge tool looks for existing PRs for its merges and stops if it
finds any. However, the 'head user' specified by the GH API for listing
pull requests is not, in fact, the user name of the person who created
the PR, but the repo owner of the repo the branch was created on. The
code has been changed accordingly.