Create destination directory in DownloadFile
On a clean enlistment (such as in an official build), one of the first
things that happens is downloading the ProdCon version-override file. If
the destination directory (`artifacts/toolset`) doesn't exist, the overall
build will fail when DownloadFile's destination directory doesn't exist.

Create it instead. If it already exists, `Directory.CreateDirectory` does
not error so this should be safe.