Replace special casing of feature branches (#302)
Feature branches currently have a special rule that treats any
test failure as a blocker. This is because many feature branches
don't have required tests, so otherwise all test failures would be
ignored. Rather than special casing feature branches, it's better
to recognize when there are no required branches and treat that
as the signifier instead