Remove post sign validation
The post sign validation phase had very questionable value in the
SignTool. The intent was to verify assemblies were actually signed by
the sign process by calling into the CLR to verify their strong name
signature. That is flawed for a couple of reasons:

1. Strong naming is only part of the verification process hence the
verification is incomplete.
1. The exit code of signing actually tells us if signing succeeded or
not. Hence doing this validation is essentially not trusting the tooling
return code which is probably over paranoid.

Additionally this was causing problems when signing native binaries. To
the sign tool it's impossible to tell whether a.exe is a managed or
native binary. Strong name validation will always fail for a.exe even
when it is properly cert signed. Hence simplyfiying things here by
removing the check altogether.