Update config.xml
This change removes the configuration to automatically create PRs from release branches to `main` in the dotnet/project-system repo. The repo doesn't tend to take many changes to release branches, and most of them need to be validated in `main` first anyway. We also sometimes make changes specific to a particular release branch that _shouldn't_ merge back to `main`.